+++ tests/PhpWord/Tests/Element/TextRunTest.php
@@ -99,7 +99,7 @@ class TextRunTest extends \PHPUnit_Framework_TestCase
 
         $this->assertInstanceOf('PhpOffice\\PhpWord\\Element\\Link', $element);
         $this->assertCount(1, $oTextRun->getElements());
-        $this->assertEquals($element->getTarget(), 'http://www.google.fr');
+        $this->assertEquals($element->getSource(), 'http://www.google.fr');
     }
 
     /**
@@ -112,7 +112,7 @@ class TextRunTest extends \PHPUnit_Framework_TestCase
 
         $this->assertInstanceOf('PhpOffice\\PhpWord\\Element\\Link', $element);
         $this->assertCount(1, $oTextRun->getElements());
-        $this->assertEquals($element->getTarget(), 'http://www.google.fr');
+        $this->assertEquals($element->getSource(), 'http://www.google.fr');
         $this->assertEquals($element->getText(), 'ééé');
     }